完成了数据库的优化修改，将课程对应的教学计划id放到course表中，解决了多个教学计划有相同的课程而被覆盖的情况。此外完成了个人修课信息的筛选和测试。

+++ src/main/java/com/scoreanalysis/bean/StuCourse.java
@@ -7,7 +7,7 @@ public class StuCourse {
 
     private String cid;
 
-    private Integer score;
+    private Double score;
 
     public String getScid() {
         return scid;
@@ -33,11 +33,11 @@ public class StuCourse {
         this.cid = cid;
     }
 
-    public Integer getScore() {
+    public Double getScore() {
         return score;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(Double score) {
         this.score = score;
     }
 
@@ -48,7 +48,7 @@ public class StuCourse {
 
         StuCourse stuCourse = (StuCourse) o;
 
-//        if (scid != null ? !scid.equals(stuCourse.scid) : stuCourse.scid != null) return false; //uuid生成，无可比性
+//        if (scid != null ? !scid.equals(stuCourse.scid) : stuCourse.scid != null) return false;
         if (sid != null ? !sid.equals(stuCourse.sid) : stuCourse.sid != null) return false;
         if (cid != null ? !cid.equals(stuCourse.cid) : stuCourse.cid != null) return false;
         return score != null ? score.equals(stuCourse.score) : stuCourse.score == null;
